:tractor: relocated definitions to core folder
