+++ moban/mobanfile/__init__.py
@@ -9,13 +9,13 @@ from lml.utils import do_import
 from moban.utils import verify_the_existence_of_directories
 from moban.externals import reporter
 from moban.deprecated import handle_copy, handle_requires
+from moban.core.data_loader import merge
 from moban.mobanfile.targets import (
     parse_targets,
     extract_target,
     extract_group_targets,
 )
 from moban.core.moban_factory import expand_template_directories
-from moban.core.data_loader import merge
 
 LOG = logging.getLogger(__name__)
